swift mailer

+++ components/cache/Memcache.php
@@ -47,16 +47,20 @@ class Memcache
 
     public function get($key)
     {
-        return $this->conn->get($key);
+        return $this->conn ? $this->conn->get($key) : '';
     }
 
     public function set($key, $value, $expiration = null)
     {
-        $this->conn->add($key, $value, $expiration);
+        if ($this->conn) {
+            $this->conn->add($key, $value, $expiration);
+        }
     }
 
     public function delete($key)
     {
-        $this->conn->delete($key);
+        if ($this->conn) {
+            $this->conn->delete($key);
+        }
     }
 }
